+++ src/meta.ts
@@ -5,4 +5,5 @@ export const repondEventsMeta = {
   defaultElapsedTimePath: null as null | StatePath<any>,
   allEventTypeGroups: {} as Record<string, Record<string, any>>,
   defaultChainId: null as null | ChainId,
+  emojiKeys: {} as Record<string, string>,
 };